メーカー関連機能追加

+++ ProductManager/src/main/java/com/example/demo/repository/ManagerRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.entity.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ManagerRepository extends JpaRepository<Manager, Integer> {
+public interface ManagerRepository extends JpaRepository<Manager, Long> {
     Manager findByEmail(String email);
 }
